+++ benches/rust_api_benchmarking.rs
@@ -1,5 +1,3 @@
-mod common;
-
 #[derive(Clone, Copy)]
 struct RustRuleBenchCase {
     n_facts_input: Option<usize>,
@@ -22,8 +20,6 @@ struct RustRuleBenchInput {
 // without doing any actual work in the rule body.
 fn match_only_rust_rule_setup(case: RustRuleBenchCase) -> RustRuleBenchInput {
     use egglog::prelude::*;
-
-    common::configure_rayon_once();
 
     let mut program = String::new();
     program.push_str("(relation R (i64))\n");
@@ -120,8 +116,6 @@ fn rust_rule_match_with_serialize(bencher: divan::Bencher, case: RustRuleBenchCa
 fn insert_loop_setup(case: RustRuleInsertLoopBenchCase) -> RustRuleBenchInput {
     use egglog::prelude::*;
 
-    common::configure_rayon_once();
-
     let mut program = String::new();
     program.push_str("(relation R (i64))\n");
     program.push_str("(function f (i64) i64 :no-merge)\n");
@@ -161,8 +155,6 @@ fn insert_loop_setup(case: RustRuleInsertLoopBenchCase) -> RustRuleBenchInput {
 
 fn tableaction_hot_path_setup(case: RustRuleTableActionBenchCase) -> RustRuleBenchInput {
     use egglog::prelude::*;
-
-    common::configure_rayon_once();
 
     let mut program = String::new();
     program.push_str("(relation R (i64))\n");
@@ -299,14 +291,12 @@ impl std::fmt::Display for ReadScanBenchCase {
 fn read_scan_setup(case: ReadScanBenchCase) -> egglog::EGraph {
     use std::fmt::Write;
 
-    common::configure_rayon_once();
-
     let mut program = String::from("(sort Math)\n(constructor Add (i64 i64) Math)\n");
     for i in 0..case.n_enodes {
         let _ = writeln!(&mut program, "(Add {} {})", i as i64, (i + 1) as i64);
     }
 
-    let mut egraph = egglog::EGraph::default();
+    let mut egraph = egglog::EGraph::new(1);
     egraph.parse_and_run_program(None, &program).unwrap();
     egraph
 }
@@ -340,8 +330,6 @@ fn main() {
 
 fn fib_setup() -> RustRuleBenchInput {
     use egglog::prelude::*;
-    common::configure_rayon_once();
-
     let mut program = String::new();
     program.push_str("(function fib (i64) i64 :no-merge)");
     program.push_str("(set (fib 0) 0)\n");